Add sort-glyphs command to CLI
Introduce functionality for sorting glyphs using various methods. The new command supports sorting by Unicode, alphabetical, or design-friendly order.
